+++ app/controllers/alaveteli_pro/plans_controller.rb
@@ -6,6 +6,10 @@ class AlaveteliPro::PlansController < AlaveteliPro::BaseController
   before_action :authenticate, :check_existing_subscription, only: [:show]
 
   def index
+    default_plan_name = add_stripe_namespace('pro')
+    stripe_plan = Stripe::Plan.retrieve(default_plan_name)
+    @plan = AlaveteliPro::WithTax.new(stripe_plan)
+    @pro_site_name = AlaveteliConfiguration.pro_site_name
   end
 
   def show