+++ app/helpers/currency_helper.rb
@@ -1,5 +1,6 @@
 module CurrencyHelper
-  def format_currency(amount)
-    Money.new(amount, AlaveteliConfiguration.iso_currency_code).format
+  def format_currency(amount, no_cents_if_whole: false)
+    Money.new(amount, AlaveteliConfiguration.iso_currency_code).
+      format(no_cents_if_whole: no_cents_if_whole)
   end
 end